Move helper hooks close to the component
