+++ site/src/utils/templateVersion.ts
@@ -1,30 +1,6 @@
 import * as API from "api/api";
 import { FileTree, createFile } from "./filetree";
 import { TarReader } from "./tar";
-import { TemplateVersion } from "api/typesGenerated";
-
-export const getTemplateFilesWithDiff = async (
-  templateName: string,
-  version: TemplateVersion,
-) => {
-  const previousVersion = await API.getPreviousTemplateVersionByName(
-    version.organization_id!,
-    templateName,
-    version.name,
-  );
-  const loadFilesPromises: ReturnType<typeof getTemplateVersionFiles>[] = [];
-  loadFilesPromises.push(getTemplateVersionFiles(version.job.file_id));
-  if (previousVersion) {
-    loadFilesPromises.push(
-      getTemplateVersionFiles(previousVersion.job.file_id),
-    );
-  }
-  const [currentFiles, previousFiles] = await Promise.all(loadFilesPromises);
-  return {
-    currentFiles,
-    previousFiles,
-  };
-};
 
 // Content by filename
 export type TemplateVersionFiles = Record<string, string>;
